pyproject.toml
[project]
name = "jobsignals"
version = "1.0.0"
description = "Job market intelligence data product"
requires-python = ">=3.11"

dependencies = [
    # Database
    "psycopg2-binary>=2.9",
    "sqlalchemy>=2.0",

    # API
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",

    # Data / validation
    "pydantic>=2.7",
    "pydantic-settings>=2.3",

    # Seed data generation
    "faker>=25.0",

    # Scheduler
    "schedule>=1.2",

    # LLM
    "anthropic>=0.40",

    # Utilities
    "python-dotenv>=1.0",
    "rich>=13.7",          # pretty CLI output
    "click>=8.1",          # CLI entry points
    "httpx>=0.27",         # for future scraper HTTP calls
    "tenacity>=8.3",       # retry logic
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]

[project.scripts]
jobsignals-api    = "api.main:start"
jobsignals-seed   = "scripts.seed_data:main"
jobsignals-ingest = "scripts.run_pipeline:ingest"
jobsignals-sched  = "scripts.scheduler:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"
